refactor: Move pagination helper functions from `core` to `ktor-utils`
This makes them available to components.
