+++ core/src/main/kotlin/api/UserWithGroupsHelper.kt
@@ -21,12 +21,12 @@ package org.eclipse.apoapsis.ortserver.core.api
 
 import org.eclipse.apoapsis.ortserver.api.v1.mapping.mapToApi
 import org.eclipse.apoapsis.ortserver.api.v1.model.UserWithGroups
-import org.eclipse.apoapsis.ortserver.core.utils.paginate
 import org.eclipse.apoapsis.ortserver.dao.QueryParametersException
 import org.eclipse.apoapsis.ortserver.model.User
 import org.eclipse.apoapsis.ortserver.model.UserGroup
 import org.eclipse.apoapsis.ortserver.shared.apimodel.PagingOptions
 import org.eclipse.apoapsis.ortserver.shared.apimodel.SortDirection
+import org.eclipse.apoapsis.ortserver.shared.ktorutils.paginate
 
 /**
  * Sort and paginate the list of [UserWithGroups] by the given [PagingOptions].
